+++ tests/api/testEvent.py
@@ -30,7 +30,8 @@ class TesteEventProcess(TestCase):
     def test_success(self):
         event = starkbank.event.parse(
             content=self.content,
-            signature=self.valid_signature
+            signature=self.valid_signature,
+            environment=exampleProject.environment,
         )
 
         print(event)
@@ -40,6 +41,7 @@ class TesteEventProcess(TestCase):
             starkbank.event.parse(
                 content=self.content,
                 signature=self.invalid_signature,
+                environment=exampleProject.environment,
             )
 
     def test_malformed_signature(self):
@@ -47,6 +49,7 @@ class TesteEventProcess(TestCase):
             starkbank.event.parse(
                 content=self.content,
                 signature=self.malformed_signature,
+                environment=exampleProject.environment,
             )