+++ CHANGELOG.md
@@ -15,6 +15,7 @@ Given a version number MAJOR.MINOR.PATCH, increment:
 ## [Unreleased]
 ### Add
 - transfer.scheduled parameter to allow Transfer scheduling
+- Transaction query by tags
 
 ## [0.4.0] - 2020-06-03
 ### Add